Begin deprecation of py.path usage
py.path has been deprecated for a long time, and since we moved to
Python 3 there is no reason now to not use the standard library
pathlib.Path objects.

This commit enables a DeprecationWarning when the pytest fixture
does not specify a preferred return object type.
A deprecation plan is laid out in HISTORY.rst